+++ src/modules/ThatPasteButton.cpp
@@ -7,8 +7,11 @@ void ThatPasteButton::onEditor() {
 
     if (getSetting<bool, "toggle-paste-state">()) {
         if (auto pasteStateButton = typeinfo_cast<CCMenuItemSpriteExtra*>(buttons->getChildByID("paste-state-button"))) {
-            tinker::utils::hijackButton(pasteStateButton, [this](auto orig, auto sender) {
-                if ((getSetting<bool, "always-show">() && m_editorUI->m_selectedObject) || (!getSetting<bool, "always-show">() && m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1)) {
+            tinker::utils::hijackButton(pasteStateButton, [this, pasteStateButton] (auto orig, auto sender) {
+                bool alwaysShow = getSetting<bool, "always-show">();
+                bool hasMultiple = m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1;
+
+                if (((alwaysShow && m_editorUI->m_selectedObject || hasMultiple) || (!alwaysShow && hasMultiple)) && pasteStateButton->getOpacity() == 255) {
                     createQuickPopup("Paste State?", "Pasting state is <cr>dangerous</c>! Are you sure?", "Cancel", "Yes", [this, orig, sender] (FLAlertLayer*, bool yes) {
                         if (yes) orig(sender);
                     });
@@ -20,8 +23,11 @@ void ThatPasteButton::onEditor() {
 
     if (getSetting<bool, "toggle-paste-color">()) {
         if (auto pasteColorButton = typeinfo_cast<CCMenuItemSpriteExtra*>(buttons->getChildByID("paste-color-button"))) {
-            tinker::utils::hijackButton(pasteColorButton, [this](auto orig, auto sender) {
-                if ((getSetting<bool, "always-show">() && m_editorUI->m_selectedObject) || (!getSetting<bool, "always-show">() && m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1)) {
+            tinker::utils::hijackButton(pasteColorButton, [this, pasteColorButton] (auto orig, auto sender) {
+                bool alwaysShow = getSetting<bool, "always-show">();
+                bool hasMultiple = m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1;
+                
+                if (((alwaysShow && m_editorUI->m_selectedObject || hasMultiple) || (!alwaysShow && hasMultiple)) && pasteColorButton->getOpacity() == 255) {
                     createQuickPopup("Paste Color?", "Pasting color is <cr>dangerous</c>! Are you sure?", "Cancel", "Yes", [this, orig, sender] (FLAlertLayer*, bool yes) {
                         if (yes) orig(sender);
                     });
@@ -41,8 +47,11 @@ void ThatPasteButton::onSetGroupIDLayer(SetGroupIDLayer* setGroupIDLayer, GameOb
     auto pasteButton = typeinfo_cast<CCMenuItemSpriteExtra*>(actions->getChildByID("paste-button"));
     if (!pasteButton) return;
 
-    tinker::utils::hijackButton(pasteButton, [this](auto orig, auto sender) {
-        if ((getSetting<bool, "always-show">() && m_editorUI->m_selectedObject) || (!getSetting<bool, "always-show">() && m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1)) {
+    tinker::utils::hijackButton(pasteButton, [this] (auto orig, auto sender) {
+        bool alwaysShow = getSetting<bool, "always-show">();
+        bool hasMultiple = m_editorUI->m_selectedObjects && m_editorUI->m_selectedObjects->count() > 1;
+                
+        if (((alwaysShow && m_editorUI->m_selectedObject || hasMultiple) || (!alwaysShow && hasMultiple))) {
             createQuickPopup("Paste State?", "Pasting state is <cr>dangerous</c>! Are you sure?", "Cancel", "Yes", [this, orig, sender] (FLAlertLayer*, bool yes) {
                 if (yes) orig(sender);
             });